Edited authenticateLocation
Location has an additional attribute, live_period, that indicates to telegram that it is a live location instead of a static location. The bot will now only accept live locations from telegram. However, the keyboard for sending location (under bookIn) has not been edited and will not work for live location.